requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.3
Django==5.2
django-axes==7.1.0
django-crispy-forms==2.4
django-encrypted-model-fields==0.6.5
django-environ==0.12.0
django-extensions==4.1
djangorestframework==3.16.0
exceptiongroup==1.2.2
iniconfig==2.1.0
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
pyOpenSSL==25.0.0
pytest==8.3.5
pytest-django==4.11.1
python-dotenv==1.1.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
Werkzeug==3.1.3
bcrypt>=3.1.7
